pyproject.toml
[project]
name = "ibms"
version = "2.12.0"
description = "Integrated Business Management System corporate application"
authors = [{ name = "Ashley Felton", email = "ashley.felton@dbca.wa.gov.au" }]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
  "django==5.2.13",
  "psycopg[binary,pool]==3.3.3",
  "python-dotenv==1.2.2",
  "dj-database-url==3.1.2",
  "gunicorn[fast]==25.3.0",
  "django-extensions==4.1",
  "django-crispy-forms==2.6",
  "crispy-bootstrap5==2026.3",
  "webtemplate-dbca==1.8.3",
  "whitenoise[brotli]==6.12.0",
  "openpyxl==3.1.5",
  # TODO: deprecate xl* libs
  "xlrd==2.0.2",
  "xlwt==1.3.0",
  "xlutils==2.0.0",
  "django-reversion==6.1.0",
  "markupsafe>=3.0.3",
  "dbca-utils==2.1.5",
  "azure-storage-blob==12.28.0",
  "django-tasks==0.12.0",
  "django-tasks-db==0.12.0",
]

[dependency-groups]
dev = [
  "ipython>=9.12.0",
  "ipdb>=0.13.13",
  "mixer>=7.2.2",
  "pre-commit>=4.5.1",
  "coverage>=7.13.0",
]

# Reference: https://docs.astral.sh/ruff/configuration/
[tool.ruff]
line-length = 140
indent-width = 4

[tool.ruff.lint]
ignore = [
  "E501", # Line too long
  "E722", # Bare except
]

# Reference: https://www.djlint.com/docs/configuration/
[tool.djlint]
profile = "django"
